refactor: prepare stats extension
